services/experiments/pyproject.toml
[project]
name = "experiements"
version = "0.1.0"
description = "This is for testing out different experiemental apps"
authors = [
    { name = "Lulzx", email = "wiserishabh@gmail.com" }
]
dependencies = [
    "streamlit>=1.33.0",
    "pandas>=2.2.2",
    "ollama>=0.2.0",
    "openai>=1.42.0",
    "pymupdf>=1.24.9",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.25.1",
    "tabulate>=0.9.0",
    "python-dotenv>=1.0.1",
    "safetensors>=0.4.4",
    "lmql>=0.7.3",
    "transformers>=4.44.2",
    "nltk>=3.9.1",
    "numpy>=2.1.0",
    "rouge-metric>=1.0.1",
    "sentence-transformers>=3.0.1",
    "torch>=2.4.0",
    "bert-score>=0.3.13",
    "rouge-score>=0.1.2",
    "thinc>=9.0.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/web"]
